+++ tests/conftest.py
@@ -1,6 +1,13 @@
+import os
 from pathlib import Path
 
 import pytest
+
+
+@pytest.fixture(autouse=True)
+def reset_environment() -> None:
+    if "HUBSPOT_ACCESS_TOKEN" in os.environ:
+        del os.environ["HUBSPOT_ACCESS_TOKEN"]
 
 
 @pytest.fixture