+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Add initial converter implementation
 - HTML: Add addons for custom code blocks, and header links
 - Add linkcheck subsystem
+- Add HubSpot API subsystem and "upload" feature
 
 
 ## 2023/09/xx 0.0.0